Add getWeatherData middleware
The weather data needs to be fetched each time a user opens the dashboard (/ route). Solution is a custom middleware that gets called immediately after user is authenticated.

In order to limit excessive calls to the weather API, checks are implemented in getWeatherData middleware. This is combined with storing the weather data and the datetime weather data was last fetched all in the DB on the user document.

Index view template (index.ejs) updated to show basic weather information for each date.