REFACTOR class names for tasks and todos
The class name responsible for hiding or showing the tasks and todos was coupled to the id field of the tasks and todos objects. Changes to the id value in the DB would cause issues in this feature. Decoupled this field by changing the class name to be generated from the index position of task or todo object.